+++ server/src/main/java/org/opensearch/common/settings/ClusterSettings.java
@@ -134,6 +134,8 @@ import org.opensearch.node.resource.tracker.ResourceTrackerSettings;
 import org.opensearch.persistent.PersistentTasksClusterService;
 import org.opensearch.persistent.decider.EnableAssignmentDecider;
 import org.opensearch.plugins.PluginsService;
+import org.opensearch.ratelimitting.admissioncontrol.AdmissionControlSettings;
+import org.opensearch.ratelimitting.admissioncontrol.settings.CPUBasedAdmissionControllerSettings;
 import org.opensearch.repositories.fs.FsRepository;
 import org.opensearch.rest.BaseRestHandler;
 import org.opensearch.script.ScriptService;
@@ -153,8 +155,6 @@ import org.opensearch.tasks.TaskResourceTrackingService;
 import org.opensearch.tasks.consumer.TopNSearchTasksLogger;
 import org.opensearch.telemetry.TelemetrySettings;
 import org.opensearch.threadpool.ThreadPool;
-import org.opensearch.throttling.admissioncontrol.AdmissionControlSettings;
-import org.opensearch.throttling.admissioncontrol.settings.IOBasedAdmissionControllerSettings;
 import org.opensearch.transport.ProxyConnectionStrategy;
 import org.opensearch.transport.RemoteClusterService;
 import org.opensearch.transport.RemoteConnectionStrategy;
@@ -685,10 +685,9 @@ public final class ClusterSettings extends AbstractScopedSettings {
                 IndicesService.CLUSTER_REMOTE_TRANSLOG_BUFFER_INTERVAL_SETTING,
                 IndicesService.CLUSTER_REMOTE_INDEX_RESTRICT_ASYNC_DURABILITY_SETTING,
                 AdmissionControlSettings.ADMISSION_CONTROL_TRANSPORT_LAYER_MODE,
-                IOBasedAdmissionControllerSettings.IO_BASED_ADMISSION_CONTROLLER_TRANSPORT_LAYER_MODE,
-                IOBasedAdmissionControllerSettings.GLOBAL_IO_USAGE_AC_LIMIT_SETTING,
-                IOBasedAdmissionControllerSettings.IO_USAGE_SEARCH_AC_LIMIT_SETTING,
-                IOBasedAdmissionControllerSettings.IO_USAGE_INDEXING_AC_LIMIT_SETTING
+                CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER_TRANSPORT_LAYER_MODE,
+                CPUBasedAdmissionControllerSettings.GLOBAL_CPU_USAGE_INDEXING_AC_LIMIT,
+                CPUBasedAdmissionControllerSettings.GLOBAL_CPU_USAGE_SEARCH_AC_LIMIT
             )
         )
     );
