+++ server/src/main/java/org/opensearch/node/Node.java
@@ -196,6 +196,8 @@ import org.opensearch.plugins.SearchPipelinePlugin;
 import org.opensearch.plugins.SearchPlugin;
 import org.opensearch.plugins.SystemIndexPlugin;
 import org.opensearch.plugins.TelemetryPlugin;
+import org.opensearch.ratelimitting.admissioncontrol.AdmissionControlService;
+import org.opensearch.ratelimitting.admissioncontrol.transport.AdmissionControlTransportInterceptor;
 import org.opensearch.repositories.RepositoriesModule;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.rest.RestController;
@@ -234,13 +236,11 @@ import org.opensearch.telemetry.tracing.TracerFactory;
 import org.opensearch.threadpool.ExecutorBuilder;
 import org.opensearch.threadpool.RunnableTaskExecutionListener;
 import org.opensearch.threadpool.ThreadPool;
-import org.opensearch.throttling.admissioncontrol.AdmissionControlService;
-import org.opensearch.throttling.admissioncontrol.transport.AdmissionControlTransportInterceptor;
-import org.opensearch.transport.TransportInterceptor;
-import org.opensearch.transport.TransportService;
-import org.opensearch.transport.Transport;
-import org.opensearch.transport.TransportInterceptorRegistry;
 import org.opensearch.transport.RemoteClusterService;
+import org.opensearch.transport.Transport;
+import org.opensearch.transport.TransportInterceptor;
+import org.opensearch.transport.TransportInterceptorRegistry;
+import org.opensearch.transport.TransportService;
 import org.opensearch.usage.UsageService;
 import org.opensearch.watcher.ResourceWatcherService;
 
@@ -824,13 +824,15 @@ public class Node implements Closeable {
             );
             TransportInterceptorRegistry transportInterceptorRegistry = new TransportInterceptorRegistry();
 
-            final AdmissionControlService admissionControlService = AdmissionControlService.newAdmissionControlService(
+            final AdmissionControlService admissionControlService = new AdmissionControlService(
                 settings,
                 clusterService.getClusterSettings(),
                 threadPool
             );
 
-            AdmissionControlTransportInterceptor admissionControlTransportInterceptor = new AdmissionControlTransportInterceptor(admissionControlService);
+            AdmissionControlTransportInterceptor admissionControlTransportInterceptor = new AdmissionControlTransportInterceptor(
+                admissionControlService
+            );
             transportInterceptorRegistry.registerTransportInterceptor(admissionControlTransportInterceptor);
 
             final AliasValidator aliasValidator = new AliasValidator();
