+++ server/src/main/java/org/opensearch/ratelimitting/admissioncontrol/settings/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+/**
+ * This package contains settings related classes for the different admission controllers
+ */
+package org.opensearch.ratelimitting.admissioncontrol.settings;
