+++ server/src/main/java/org/opensearch/ratelimitting/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Base OpenSearch Throttling package
+ */
+package org.opensearch.ratelimitting;
