--- server/src/main/java/org/opensearch/throttling/admissioncontrol/controllers/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * This package contains classes related to the different admission controllers
- */
-package org.opensearch.throttling.admissioncontrol.controllers;
