--- server/src/main/java/org/opensearch/throttling/admissioncontrol/enums/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * This package contains enums related to the different admission controller feature
- */
-package org.opensearch.throttling.admissioncontrol.enums;
