--- server/src/main/java/org/opensearch/throttling/admissioncontrol/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * This package contains base classes needed for the admissionController Feature
- */
-package org.opensearch.throttling.admissioncontrol;
