+++ server/src/test/java/org/opensearch/ratelimitting/admissioncontrol/transport/AdmissionControlTransportHandlerTests.java
@@ -1,0 +1,93 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.ratelimitting.admissioncontrol.transport;
+
+import org.opensearch.core.concurrency.OpenSearchRejectedExecutionException;
+import org.opensearch.ratelimitting.admissioncontrol.AdmissionControlService;
+import org.opensearch.tasks.Task;
+import org.opensearch.test.OpenSearchTestCase;
+import org.opensearch.transport.TransportChannel;
+import org.opensearch.transport.TransportRequest;
+import org.opensearch.transport.TransportRequestHandler;
+
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class AdmissionControlTransportHandlerTests extends OpenSearchTestCase {
+    AdmissionControlTransportHandler<TransportRequest> admissionControlTransportHandler;
+    private String testAction = "";
+
+    public void testHandlerInvoked() throws Exception {
+        String action = "TEST";
+        InterceptingRequestHandler<TransportRequest> handler = new InterceptingRequestHandler<>(action);
+        admissionControlTransportHandler = new AdmissionControlTransportHandler<TransportRequest>(
+            action,
+            handler,
+            mock(AdmissionControlService.class),
+            false
+        );
+        admissionControlTransportHandler.messageReceived(mock(TransportRequest.class), mock(TransportChannel.class), mock(Task.class));
+        assertEquals(1, handler.count);
+    }
+
+    public void testHandlerInvokedRejectedException() throws Exception {
+        String action = "TEST";
+        AdmissionControlService admissionControlService = mock(AdmissionControlService.class);
+        when(admissionControlService.applyTransportAdmissionControl(testAction)).thenThrow(new OpenSearchRejectedExecutionException());
+        InterceptingRequestHandler<TransportRequest> handler = new InterceptingRequestHandler<>(action);
+        admissionControlTransportHandler = new AdmissionControlTransportHandler<TransportRequest>(
+            action,
+            handler,
+            admissionControlService,
+            false
+        );
+        try {
+            admissionControlTransportHandler.messageReceived(mock(TransportRequest.class), mock(TransportChannel.class), mock(Task.class));
+        } catch (OpenSearchRejectedExecutionException exception) {
+            assertEquals(0, handler.count);
+            handler.messageReceived(mock(TransportRequest.class), mock(TransportChannel.class), mock(Task.class));
+        }
+        assertEquals(1, handler.count);
+    }
+
+    public void testHandlerInvokedRandomException() throws Exception {
+        String action = "TEST";
+        AdmissionControlService admissionControlService = mock(AdmissionControlService.class);
+        when(admissionControlService.applyTransportAdmissionControl(testAction)).thenThrow(new NullPointerException());
+        InterceptingRequestHandler<TransportRequest> handler = new InterceptingRequestHandler<>(action);
+        admissionControlTransportHandler = new AdmissionControlTransportHandler<TransportRequest>(
+            action,
+            handler,
+            admissionControlService,
+            false
+        );
+        try {
+            admissionControlTransportHandler.messageReceived(mock(TransportRequest.class), mock(TransportChannel.class), mock(Task.class));
+        } catch (Exception exception) {
+            assertEquals(0, handler.count);
+            handler.messageReceived(mock(TransportRequest.class), mock(TransportChannel.class), mock(Task.class));
+        }
+        assertEquals(1, handler.count);
+    }
+
+    private class InterceptingRequestHandler<T extends TransportRequest> implements TransportRequestHandler<T> {
+        private final String action;
+        public int count;
+
+        public InterceptingRequestHandler(String action) {
+            this.action = action;
+            this.count = 0;
+        }
+
+        @Override
+        public void messageReceived(T request, TransportChannel channel, Task task) throws Exception {
+            this.count = this.count + 1;
+        }
+    }
+}
